Refactor code with lint

+++ samples/foobar/foobar.py
@@ -7,18 +7,21 @@ class Foo:
     def get_info(self) -> str:
         return "Foo Class"
 
+
 class Bar:
-    def bar_func(self, x: int) -> int:
+    def bar_func(self, x: int) -> float:
         return x / 2
 
     def get_info(self) -> str:
         return "Bar Class"
+
 
 def main():
     foo = Foo()
     bar = Bar()
     print(foo.foo_func(16))
     print(bar.bar_func(16))
+
 
 if __name__ == "__main__":
     main()
